src/llm_lib/requirements.txt
pydantic
openai
python-dotenv
tiktoken